requirements.txt
# ---- Core LiveKit Agent ----
livekit-agents>=0.10.0
livekit>=0.17.0

# ---- AI Plugins ----
livekit-plugins-openai>=0.9.0
livekit-plugins-simli>=1.2.0
livekit-plugins-sarvam>=0.1.0
livekit-plugins-google>=0.1.0
livekit-plugins-anam>=0.1.0
livekit-plugins-groq>=0.1.0
livekit-plugins-bey>=0.1.0
openpyxl>=0.1.0

# ---- Utilities ----
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
aiohttp>=3.9.0
lxml>=5.0.0
flask>=3.0.0
flask-cors>=4.0.0
